requirements.txt
flask
python-dotenv
bcrypt
